Enable Emacs fullscreem at startup
I disabled it to see my Menu Bar Pomodoro Timer when working on Emacs.

But I realized that I can see it just by moving my mouse pointer to the
top. I don't need to switch windows.

Besides, my timer already has a short alarm when the Pomodoro is about
to finish, to warn me to start to wrap up the current Pomodoro.